#10070: Move cosh bw op to TTNN
