+++ docs/source/ttnn/ttnn/dependencies/tt_lib.rst
@@ -838,8 +838,6 @@ Backward Operations
 
 .. autofunction:: tt_lib.tensor.polygamma_bw
 
-.. autofunction:: tt_lib.tensor.cosh_bw
-
 .. autofunction:: tt_lib.tensor.erfinv_bw
 
 .. autofunction:: tt_lib.tensor.hardtanh_bw
