+++ docs/source/ttnn/ttnn/ttnn/cosh_bw.rst
@@ -1,0 +1,6 @@
+.. _ttnn.cosh_bw:
+
+ttnn.cosh_bw
+###############
+
+.. autofunction:: ttnn.cosh_bw
