+++ tests/ttnn/unit_tests/operations/backward/test_backward_cosh.py
@@ -4,8 +4,8 @@
 
 import torch
 import pytest
-import tt_lib
-from tests.tt_eager.python_api_testing.unit_testing.backward_ops.utility_funcs import (
+import ttnn
+from tests.ttnn.unit_tests.operations.backward.utility_funcs import (
     data_gen_with_range,
     compare_pcc,
 )
@@ -26,7 +26,7 @@ def test_bw_cosh(input_shapes, device):
     in_data, input_tensor = data_gen_with_range(input_shapes, -9, 9, device, True)
     grad_data, grad_tensor = data_gen_with_range(input_shapes, -20, 20, device)
 
-    tt_output_tensor_on_device = tt_lib.tensor.cosh_bw(grad_tensor, input_tensor)
+    tt_output_tensor_on_device = ttnn.cosh_bw(grad_tensor, input_tensor)
 
     in_data.retain_grad()
 
@@ -52,7 +52,7 @@ def test_bw_cosh_inf(input_shapes, device):
     in_data, input_tensor = data_gen_with_range(input_shapes, 90, 95, device, True)
     grad_data, grad_tensor = data_gen_with_range(input_shapes, -7, 7, device)
 
-    tt_output_tensor_on_device = tt_lib.tensor.cosh_bw(grad_tensor, input_tensor)
+    tt_output_tensor_on_device = ttnn.cosh_bw(grad_tensor, input_tensor)
 
     in_data.retain_grad()
 
@@ -78,7 +78,7 @@ def test_bw_cosh_neg_inf(input_shapes, device):
     in_data, input_tensor = data_gen_with_range(input_shapes, -95, -89, device, True)
     grad_data, grad_tensor = data_gen_with_range(input_shapes, -7, 7, device)
 
-    tt_output_tensor_on_device = tt_lib.tensor.cosh_bw(grad_tensor, input_tensor)
+    tt_output_tensor_on_device = ttnn.cosh_bw(grad_tensor, input_tensor)
 
     in_data.retain_grad()
 
@@ -101,7 +101,7 @@ def test_bw_cosh_nan_test1(input_shapes, device):
     in_data, input_tensor = data_gen_with_range(input_shapes, 86, 89, device, True)
     grad_data, grad_tensor = data_gen_with_range(input_shapes, 35, 50, device)
 
-    tt_output_tensor_on_device = tt_lib.tensor.cosh_bw(grad_tensor, input_tensor)
+    tt_output_tensor_on_device = ttnn.cosh_bw(grad_tensor, input_tensor)
 
     in_data.retain_grad()
 
@@ -124,7 +124,7 @@ def test_bw_cosh_nan_test2(input_shapes, device):
     in_data, input_tensor = data_gen_with_range(input_shapes, 86, 89, device, True)
     grad_data, grad_tensor = data_gen_with_range(input_shapes, -50, -35, device)
 
-    tt_output_tensor_on_device = tt_lib.tensor.cosh_bw(grad_tensor, input_tensor)
+    tt_output_tensor_on_device = ttnn.cosh_bw(grad_tensor, input_tensor)
 
     in_data.retain_grad()
