+++ tt_eager/tt_dnn/op_library/backward/backward_ops.cpp
@@ -421,39 +421,6 @@ std::vector<Tensor> polygamma_bw(
     return operation::decorate_as_composite(__func__, _polygamma_bw)(grad, input, n, output_mem_config);
 }
 
-// name: cosh(Tensor self) -> Tensor
-// self: grad * self.sinh()
-std::vector<Tensor> _cosh_bw(const Tensor& grad, const Tensor& input, const MemoryConfig& output_mem_config) {
-    std::vector<Tensor> grad_tensor;
-    Tensor t_inf = ttnn::multiply(ttnn::sign(grad, output_mem_config), std::numeric_limits<float>::infinity(), std::nullopt, output_mem_config);
-    Tensor t_neg_inf =
-        ttnn::multiply(ttnn::sign(grad, output_mem_config), -std::numeric_limits<float>::infinity(), std::nullopt, output_mem_config);
-    Tensor grad_a = where(
-        ttnn::gt(input, full_like(input, 88.50, output_mem_config), std::nullopt, output_mem_config),
-        t_inf,
-        where(
-            ttnn::lt(input, full_like(input, -88.50, output_mem_config), std::nullopt, output_mem_config),
-            t_neg_inf,
-            ttnn::multiply(grad, sinh(input, output_mem_config), std::nullopt, output_mem_config),
-            output_mem_config),
-        output_mem_config);
-    t_neg_inf.deallocate();
-    t_inf.deallocate();
-    grad_a = where(
-        ttnn::ge(grad_a, 3.4e+38, std::nullopt, output_mem_config),
-        std::numeric_limits<float>::infinity(),
-        where(
-            ttnn::le(grad_a, -3.4e+38, std::nullopt, output_mem_config),
-            -std::numeric_limits<float>::infinity(),
-            grad_a,
-            output_mem_config),
-        output_mem_config);
-    grad_tensor.emplace_back(grad_a);
-    return grad_tensor;
-}
-std::vector<Tensor> cosh_bw(const Tensor& grad, const Tensor& input, const MemoryConfig& output_mem_config) {
-    return operation::decorate_as_composite(__func__, _cosh_bw)(grad, input, output_mem_config);
-}
 
 // Hardtanh
 // result: torch.where((input <= min) | (input >= max), 0.0, grad)
