+++ ttnn/cpp/ttnn/operations/eltwise/unary_backward/device/unary_backward_op.hpp
@@ -67,6 +67,7 @@ enum class UnaryBackwardOpType {
     ERFC_BW,
     CEIL_BW,
     SOFTSIGN_BW,
+    COSH_BW,
 };
 
 struct UnaryBackwardFunction{
